test(svelte): split zoom-state harness by decision group (#558)
Mirror the focus/interval harness pattern so construction, domain, and
integration coverage stay locally editable without a 700-line monolith.
